+++ evealert/__init__.py
@@ -1,4 +1,4 @@
 """Initialize the app"""
 
-__version__ = "5.0.0"
+__version__ = "5.0.1"
 __title__ = "EVE Alert"